add plugin command added

+++ src/test/java/org/walkmod/WalkmodDispatcherTest.java
@@ -109,7 +109,7 @@ public class WalkmodDispatcherTest {
 			String userDir = new File(System.getProperty("user.dir")).getAbsolutePath();
 			System.setProperty("user.dir", tmp.getAbsolutePath());
 			try {
-				run(new String[] { "add-chain", "-d", "src", "--walker = { transformations :[ { type: \"walkmod:commons:import-cleaner\"} ] }" });
+				run(new String[] { "add-chain", "-d", "src", "--walker" ,"{ transformations :[ { type: \"walkmod:commons:import-cleaner\"} ] }" });
 			}
 			finally {
 				System.setProperty("user.dir", userDir);
@@ -122,6 +122,35 @@ public class WalkmodDispatcherTest {
 				System.out.println(content);
 				
 				Assert.assertTrue(content.contains("import-cleaner"));
+				
+				
+				FileUtils.deleteDirectory(tmp);
+			}
+			
+		}
+	}
+	
+	@Test
+	public void testAddPlugin() throws Exception{
+		File tmp = new File("src/test/resources/initTestAddPlugin");
+		tmp.mkdirs();
+		if(tmp.exists()){
+			String userDir = new File(System.getProperty("user.dir")).getAbsolutePath();
+			System.setProperty("user.dir", tmp.getAbsolutePath());
+			try {
+				run(new String[] { "add-plugin", "--groupId", "org.walkmod", "--artifactId", "walkmod-imports-cleaner-plugin", "--version", "2.0" });
+			}
+			finally {
+				System.setProperty("user.dir", userDir);
+				
+				File cfg = new File(tmp, "walkmod.xml");
+				Assert.assertTrue(cfg.exists());
+				
+				String content = FileUtils.readFileToString(cfg);
+				
+				System.out.println(content);
+				
+				Assert.assertTrue(content.contains("walkmod-imports-cleaner-plugin"));
 				
 				
 				FileUtils.deleteDirectory(tmp);
